requirements/keras-common.txt
tf_keras==2.15.0
keras-nlp==0.8.2
keras-cv==0.8.2
tensorflow-datasets==4.9.3
